deps: upgrade to libuv 1.38.1
Notable changes:

- A probable compiler bug in VS2019 was causing a
  failed assertion in libuv which appeared to the
  user as though the system clock was drifting. The
  `uv_hrtime()` code on Windows has been rearranged
  to work around the issue.
- On Linux, `uv_loadavg()` attempts to read from
  `/proc/loadavg` before falling back to calling
  `sysinfo()`. This works around a bug in LXC.
- A deadlock in the Windows TTY code has been fixed.
- An issue on macOS related to monotonic clocks jumping
  back in time has been worked around.

PR-URL: https://github.com/nodejs/node/pull/34187

+++ deps/uv/src/win/internal.h
@@ -266,7 +266,7 @@ void uv__fs_poll_endgame(uv_loop_t* loop, uv_fs_poll_t* handle);
  */
 void uv__util_init(void);
 
-uint64_t uv__hrtime(double scale);
+uint64_t uv__hrtime(unsigned int scale);
 __declspec(noreturn) void uv_fatal_error(const int errorno, const char* syscall);
 int uv__getpwuid_r(uv_passwd_t* pwd);
 int uv__convert_utf16_to_utf8(const WCHAR* utf16, int utf16len, char** utf8);
